Added user reference to books

+++ backend/controllers/bookController.js
@@ -8,7 +8,8 @@ const BookController = {
             title: req.body.title,
             author: req.body.author,
             startDate: req.body.startDate,
-            endDate: req.body.endDate
+            endDate: req.body.endDate,
+            user_id: req.user.userId
         })
         try{
             if(req.file){
@@ -18,30 +19,29 @@ const BookController = {
             const newBook = await book.save();
             res.status(201).json(newBook);
         } catch(err){
-            console.log(err.message);
             res.status(400).json({ message: err.message });
         }
     },
     async get(req, res){
         const { id } = req.params;
+        const user_id = req.user.userId;
         try{
-            const book = await Book.findById(id);
+            const book = await Book.findOne({ user_id: user_id, _id: id});
             res.status(200).json(book);
         } catch(err) {
-            console.log(err.message);
             res.status(500).json({ message: err.message });
         }
     },
     async index(req, res){
         const { field, direction } = req.query;
-        console.log(req.query);
+        const user_id = req.user.userId;
         try{
             let books;
             if(field == "endDate"){
-                books = await Book.find().sort({ endDate: direction });
+                books = await Book.find({user_id}).sort({ endDate: direction });
             }
             else{
-                books = await Book.find().sort({ startDate: direction });
+                books = await Book.find({user_id}).sort({ startDate: direction });
             }
             res.json(books);
         } catch(err){
@@ -50,14 +50,15 @@ const BookController = {
     },
     async update(req, res){
         const { id } = req.params;
-        const book = await Book.findById(id);
-        if(req.body.title){
-            book.title = req.body.title;
-        }
-        if(req.body.author){
-            book.author = req.body.author;
-        }
+        const user_id = req.user.userId;
         try {
+            const book = await Book.findOne({ user_id: user_id, _id: id});
+            if(req.body.title){
+                book.title = req.body.title;
+            }
+            if(req.body.author){
+                book.author = req.body.author;
+            }
             const updatedBook = await book.save();
             res.json(updatedBook);
         } catch(err) {
@@ -65,8 +66,9 @@ const BookController = {
         }
     },
     async delete(req, res){
+        const user_id = req.user.userId;
         try{
-            await Book.findOneAndDelete({ _id: req.params.id });
+            await Book.findOneAndDelete({ user_id: user_id, _id: req.params.id });
             res.json({ message: "Deleted book" });
         } catch(err){
             res.status(500).json({ message: err.message });
